requirements.txt
streamlit
pillow
pdf2image
python-dotenv
google-generativeai
